requirements.txt
notebook
# torch==1.13.1
# torchvision==0.14.1
# torchaudio==0.13.1
git+https://github.com/openai/CLIP.git
torch-encoding==1.2.1
# opencv-python==4.7.0.68
opencv-python==4.1.2.30
imageio==2.23.0
imageio-ffmpeg==0.4.7
pytorch-lightning==1.8.1
altair==4.1.0
timm==0.6.12
tensorboardX==2.2
matplotlib==3.6.2
test-tube==0.7.5
wandb==0.13.7
h5py==3.7.0
cmake
# gym==0.22.0
open3d==0.17.0
grad-cam==1.4.6
networkx==2.8.8
mpl_point_clicker==0.3.1
ai2thor==5.0.0
rospkg==1.4.0
git+https://github.com/TaipanRex/pyvisgraph.git
gdown==4.6.0
hydra-core==1.3.1
google-cloud-speech==2.17.0
sentence_transformers==2.2.2
shapely==2.0.2
openai==1.3.7
shapely==1.7.1
websockets
aiohttp
aiohttp_middlewares